+++ src/components/Timer.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
-import { StyleSheet, View, Text, Button, AsyncStorage } from 'react-native';
+import { StyleSheet, View, Text, Button } from 'react-native';
 import { connect } from 'react-redux';
 import PercentageCircle from 'react-native-percentage-circle';
-import moment from 'moment';
 import { changeInterval, countdown } from '../actions';
+import { sayHi, setEndTimer, retrieveEndTime, removeEndTime } from '../helper';
 
 const styles = StyleSheet.create({
   container: {
@@ -25,84 +25,52 @@ const styles = StyleSheet.create({
 export class Timer extends Component {
   constructor(props) {
     super(props);
-    this.counter = props.currentCounter;
-    this.full = props.full;
     this.timerStart = props.timerStart;
     this.changeInterval = props.changeInterval;
   }
 
   componentDidMount() {
-    // comment out for implement AsyncStorage
     const intervalID = setInterval(this.timerStart, 1000);
     this.changeInterval(intervalID);
   }
 
-  sayHi = () => {
-    console.log("Don't do that");
-    const currentTime = moment();
-    console.log(currentTime);
-    console.log(currentTime.subtract(3, 'hours').format('h:mm'));
-  };
-
-  setEndTimer = async () => {
-    try {
-      const endTime = moment().add(7, 'hours');
-      await AsyncStorage.setItem('someKey', 'someValue');
-      await AsyncStorage.setItem('endTime', endTime);
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
-  retrieveEndTime = async () => {
-    try {
-      const someKey = await AsyncStorage.getItem('someKey');
-      const endTime = await AsyncStorage.getItem('endTime');
-      const dataset = [someKey, endTime];
-
-      for (let i = 0; i < dataset.length; i++) {
-        if (dataset[i] !== null) {
-          console.log(dataset[i]);
-        }
-      }
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
-  removeEndTime = async () => {
-    try {
-      await AsyncStorage.removeItem('endTime', console.log('remove end time'));
-    } catch (err) {
-      console.log(err);
-    }
-  };
+  // shouldComponentUpdate(nextProps, nextState) {
+  //   console.log("Hi Tsuyoshi")
+  //   if (this.counter !== nextProps.currentCounter) {
+  //     return false;
+  //   }
+  //   // if (this.counter !== nextState.counter) {
+  //   //   return true;
+  //   // }
+  //   return false;
+  // }
 
   render() {
-    const percentage = `${String(Math.floor((this.counter * 100) / this.full))}%`;
-    const time = this.counter.format('h:mm');
+    const {currentCounter, full} = this.props;
+    const percentage = `${String(Math.floor((currentCounter * 100) / full))}%`;
+    const time = currentCounter.format('h:mm:ss');
     return (
       <View style={styles.container}>
         <Button
-          onPress={this.sayHi}
+          onPress={sayHi}
           title="Don't push me"
           color="#fff"
           accessibilityLabel="Don't do that"
         />
         <Button
-          onPress={this.setEndTimer}
+          onPress={setEndTimer}
           title="Save data to LS"
           color="#f6eaea"
           accessibilityLabel="Don't do that"
         />
         <Button
-          onPress={this.retrieveEndTime}
+          onPress={retrieveEndTime}
           title="Retrieve data from LS"
           color="#f6eaea"
           accessibilityLabel="Don't do that"
         />
         <Button
-          onPress={this.removeEndTime}
+          onPress={removeEndTime}
           title="Bomberman"
           color="#f6eaea"
           accessibilityLabel="Don't do that"
